Foods/Meals as Buttons, UI Adjustments, Improved Efficiency, etc.
I made it so that the foods and meals that have been added to the database are not being stored in the HTML as strings anymore. Instead, they're each being stored as a Button with the name of the food/meal as it's innerHTML. When these Buttons are clicked, they'll trigger the nutrition information of their corresponding food to display above. It looks totally rad dude!

Also made some UI adjustments. I found that a minor inconvenience was the order in which the fields were laid out in the forms. I swapped it so that the order of input fields goes name of food, unit of measurement, serving size, calories, fat, carbs, and then protein. This is the order that they appear on a nutrition label, so it makes more sense to read down the label and input as you progress downward rather than keep having your eyes jump around the label.

I also put all of the Food/Meal Buttons from the Foods and Meals page onto the Today's Nutrition page, but instead of displaying their nutrition information when clicked on, they'll prompt the user asking how many servings of that food/meal they want to log into Today's Nutrition. When it adds the nutrition, it will add it as a button, and when the button is clicked on, it can be removed from Today's Nutrition. Much faster and more efficient!